tools/gnulib: import coreutils local modules
These custom gnulib modules from coreutils source
are necessary to run automake for coreutils when using
the bootstrap script because the release tarball
distribution of coreutils is missing the "gl" directory.

Otherwise, both the freshly generated lib/gnulib.mk
and the backup lib/gnulib.mk~ are unusable in different ways.

Instructions for refreshing are in the patch.

Signed-off-by: Michael Pratt <mcpratt@pm.me>